Add day4 solution
46 for the first star and 20 for the second star.

For the first problem I started with a quick loop to find the minimal
key/value pair in the checksum. Then loop through the rest to see if
the value is smaller or, if equal, the key is smaller. It was around
60 lines so I'm happy with how compact the my compact solution ended
up.

The ceaser cipher was easy since I knew about the subtract 'a' trick.